+++ graphtransliterator/exceptions.py
@@ -1,10 +1,7 @@
 # -*- coding: utf-8 -*-
 
 """
-graphtransliterator.exceptions
-------------------------------
-
-Transliteration exceptions used in Graph Transliterator are defined here.
+Transliteration exceptions used in Graph Transliterator.
 """
 
 
@@ -18,18 +15,21 @@ class GraphTransliteratorException(Exception):
 class AmbiguousTransliterationRulesException(GraphTransliteratorException):
     """
     Raised when multiple transliteration rules can match the same pattern.
+    Details of ambiguities are given in a :func:`logging.warning`.
     """
 
 
-class NoMatchingTransliterationRule(GraphTransliteratorException):
+class NoMatchingTransliterationRuleException(GraphTransliteratorException):
     """
     Raised when no transliteration rule can be matched at a particular
-    location in the input string's tokens.
+    location in the input string's tokens. Details of the location are given
+    in a :func:`logging.warning`.
     """
 
 
-class UnrecognizableInputToken(GraphTransliteratorException):
+class UnrecognizableInputTokenException(GraphTransliteratorException):
     """
     Raised when a character in the input string does not correspond to any
-    tokens in the GraphTranslator's token settings.
+    tokens in the GraphTransliterator's token settings. Details of the location
+     are given in a :func:`logging.warning`.
     """